moved communication failed exception to diffrent namespace
