+++ TxtControllerLib/Components/Sequencer/ControllerSequencer.cs
@@ -9,7 +9,7 @@ using RoboticsTxt.Lib.Commands;
 using RoboticsTxt.Lib.Components.Communicator;
 using RoboticsTxt.Lib.Contracts;
 using RoboticsTxt.Lib.Contracts.Configuration;
-using RoboticsTxt.Lib.ControllerDriver;
+using RoboticsTxt.Lib.Contracts.Exceptions;
 
 namespace RoboticsTxt.Lib.Components.Sequencer
 {
